refactor: more visible structure in organizer
