+++ src/organizer/index.ts
@@ -1,15 +1,15 @@
 import Web3 from 'web3'
 import { logger } from '@zkopru/utils'
-import { startLogger } from './generator-utils'
-import { config } from './config'
-import { OrganizerConfig } from './types'
-import { OrganizerApi } from './organizer-api'
+import { startLogger } from '../generator-utils'
+import { config } from '../config'
+import { OrganizerConfig } from './context'
+import { Organizer } from './organizer'
 
 startLogger('ORGANIZER_LOG')
 
-logger.info('stress-test/organizer.ts - organizer initializing')
+logger.info('stress-test/organizer/index.ts - organizer initializing')
 
-const isDevelopment = process.env.DEVELOPMENT
+const isDevelopment = process.env.DEVELOPMENT ?? false
 
 const webSocketProvider = new Web3.providers.WebsocketProvider(
   isDevelopment ? 'localhost:8545' : config.testnetUrl,
@@ -18,17 +18,13 @@ const webSocketProvider = new Web3.providers.WebsocketProvider(
     timeout: 600,
   },
 )
-const web3 = new Web3(webSocketProvider)
-
-const organierContext = {
-  web3,
-  dev: isDevelopment,
-  coordinators: []
-} // Test Coordinator
 
 const organizerConfig: OrganizerConfig = {
-  connection: { host: isDevelopment ? 'localhost' : 'redis', port: 6379 },
   dev: !!isDevelopment,
+  node: {
+    redis: { host: isDevelopment ? 'localhost' : 'redis', port: 6379 },
+    web3Provider: webSocketProvider
+  },
   rates: [
     { name: '0.1', max: 1, duration: 10000 },
     { name: '1', max: 1, duration: 1000 },
@@ -40,6 +36,6 @@ const organizerConfig: OrganizerConfig = {
   organizerPort: 8080,
 }
 
-const organizer = new OrganizerApi(organierContext, organizerConfig)
+const organizer = new Organizer(organizerConfig)
 organizer.start()
-logger.info('stress-test/organizer.ts - organizer started')
+logger.info('stress-test/organizer/indext.ts - organizer started')
